+++ XNGAPIClient/XNGAPIClient+ProfileEditing.m
@@ -27,6 +27,7 @@
 - (void)putUpdateUsersGeneralInformationWithAcademicTitle:(NSString *)academicTitle
                                          employmentStatus:(NSString *)employmentStatus
                                                     haves:(NSString *)haves
+                                                 topHaves:(NSString *)topHaves
                                                 interests:(NSString *)interests
                                             organisations:(NSString *)organisations
                                                     wants:(NSString *)wants
@@ -41,6 +42,9 @@
     }
     if (haves) {
         parameters[@"haves"] = haves;
+    }
+    if (topHaves) {
+        parameters[@"top_haves"] = topHaves;
     }
     if (interests) {
         parameters[@"interests"] = interests;